+++ src/org/pepstock/charba/client/configuration/GridLines.java
@@ -212,7 +212,7 @@ public class GridLines extends AbstractScaleLines {
 	 * 
 	 * @param borderDashOffset Offset for line dashes.
 	 */
-	public void setBorderDashOffset(int borderDashOffset) {
+	public void setBorderDashOffset(double borderDashOffset) {
 		// reset callback if there is
 		setBorderDashOffset((ScaleBorderDashOffsetCallback) null);
 		// stores value
@@ -224,7 +224,7 @@ public class GridLines extends AbstractScaleLines {
 	 * 
 	 * @return Offset for line dashes.
 	 */
-	public int getBorderDashOffset() {
+	public double getBorderDashOffset() {
 		return getAxis().getScale().getGrideLines().getBorderDashOffset();
 	}
