+++ src/org/pepstock/charba/client/options/AngleLines.java
@@ -155,7 +155,7 @@ public final class AngleLines extends AbstractScaleLines<IsDefaultAngleLines> im
 	 * @see org.pepstock.charba.client.options.AbstractScaleLine#getDefaultBorderDashOffset()
 	 */
 	@Override
-	int getDefaultBorderDashOffset() {
+	double getDefaultBorderDashOffset() {
 		return getDefaultValues().getBorderDashOffset();
 	}
 }
